Added a more sophisticated mechanism for identifying a good interior point.
